Finalizando seção de ReactJS

+++ client/src/pages/books/index.js
@@ -7,27 +7,29 @@ import { FiPower, FiEdit, FiTrash2 } from 'react-icons/fi';
 
 export default function Books() {
     const [books, setBooks] = useState([]);
+    const [page, setPage] = useState(1);
     const userName = localStorage.getItem('userName');
     const accessToken = localStorage.getItem('accessToken');
+    const authorization = {
+        headers: {
+            Authorization: `Bearer ${accessToken}`
+        }
+    };
     const navigate = useNavigate();
 
     useEffect(() => {
-        api.get('Book/v1/asc/20/1', {
-            headers: {
-                Authorization: `Bearer ${accessToken}`
-            }
-        }).then(response => {
-            setBooks(response.data.list)
-        })
+        fetchMoreBooks();
     }, [accessToken]);
+
+    async function fetchMoreBooks() {
+        const response = await api.get(`Book/v1/asc/4/${page}`, authorization);
+        setBooks([...books, ...response.data.list]);
+        setPage(page + 1);
+    }
 
     async function logout() {
         try {
-            await api.get('Auth/v1/revoke', {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`
-                }
-            });
+            await api.get('Auth/v1/revoke', authorization);
 
             localStorage.clear();
             navigate('/');
@@ -37,13 +39,18 @@ export default function Books() {
         }
     }
 
+    async function editBook(id) {
+        try {
+            navigate(`/Book/new/${id}`);
+        } catch (error) {
+            alert('Edit book failed! Try again.');
+            console.log(error);
+        }
+    }
+
     async function deleteBook(id) {
         try {
-            await api.delete(`Book/v1/${id}`, {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`
-                }
-            });
+            await api.delete(`Book/v1/${id}`, authorization);
 
             setBooks(books.filter(book => book.id !== id));
         } catch (error) {
@@ -57,7 +64,7 @@ export default function Books() {
             <header>
                 <img src={logoImage} alt="Rest With ASP.NET" />
                 <span>Welcome, <strong>{userName.toUpperCase()}</strong>!</span>
-                <Link className="button" to="/newBook">Add New Book</Link>
+                <Link className="button" to="/book/new/0">Add New Book</Link>
                 <button type="button" onClick={logout}>
                     <FiPower size={18} color="#251FC5" />
                 </button>
@@ -71,10 +78,10 @@ export default function Books() {
                         <strong>Author: </strong>
                         <p>{book.author}</p>
                         <strong>Price: </strong>
-                        <p>{Intl.NumberFormat('pt-br', {style: 'currency', currency: 'BRL'}).format(book.price)}</p>
+                        <p>{Intl.NumberFormat('pt-br', { style: 'currency', currency: 'BRL' }).format(book.price)}</p>
                         <strong>Release Date: </strong>
                         <p>{Intl.DateTimeFormat('pt-br').format(new Date(book.launchDate))}</p>
-                        <button type="button">
+                        <button type="button" onClick={() => editBook(book.id)}>
                             <FiEdit size={20} color="#251FC5" />
                         </button>
                         <button type="button" onClick={() => deleteBook(book.id)}>
@@ -83,6 +90,7 @@ export default function Books() {
                     </li>
                 ))}
             </ul>
+            <button className="button" type="button" onClick={fetchMoreBooks}>Load More</button>
         </div>
     );
 }
